Added slime obstacle
The slime obstacle slows the player down by a certain factor,
and furthermore makes them jump lower by the same factor.
Note that for it to be effective, it will need to be placed slightly
above the ground, but this can be changed by extending the hitbox
vertically beyond the sprite, by a bit.